LayoutTests/platform/mac-bigsur/imported/w3c/web-platform-tests/fetch/metadata/generated/fetch.sub-expected.txt

Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY

FAIL sec-fetch-site - Not sent to non-trustworthy same-origin destination, no init assert_not_own_property: unexpected property "sec-fetch-site" is found on object
FAIL sec-fetch-site - Not sent to non-trustworthy same-site destination, no init promise_test: Unhandled rejection with value: object "TypeError: Load failed"
FAIL sec-fetch-site - Not sent to non-trustworthy cross-site destination, no init assert_not_own_property: unexpected property "sec-fetch-site" is found on object
FAIL sec-fetch-mode - Not sent to non-trustworthy same-origin destination, no init assert_not_own_property: unexpected property "sec-fetch-mode" is found on object
FAIL sec-fetch-mode - Not sent to non-trustworthy same-site destination, no init promise_test: Unhandled rejection with value: object "TypeError: Load failed"
FAIL sec-fetch-mode - Not sent to non-trustworthy cross-site destination, no init assert_not_own_property: unexpected property "sec-fetch-mode" is found on object
FAIL sec-fetch-dest - Not sent to non-trustworthy same-origin destination, no init assert_not_own_property: unexpected property "sec-fetch-dest" is found on object
FAIL sec-fetch-dest - Not sent to non-trustworthy same-site destination, no init promise_test: Unhandled rejection with value: object "TypeError: Load failed"
FAIL sec-fetch-dest - Not sent to non-trustworthy cross-site destination, no init assert_not_own_property: unexpected property "sec-fetch-dest" is found on object
PASS sec-fetch-user - Not sent to non-trustworthy same-origin destination, no init
FAIL sec-fetch-user - Not sent to non-trustworthy same-site destination, no init promise_test: Unhandled rejection with value: object "TypeError: Load failed"
PASS sec-fetch-user - Not sent to non-trustworthy cross-site destination, no init
PASS sec-fetch-site - HTTPS downgrade (header not sent), no init
FAIL sec-fetch-site - HTTPS upgrade, no init assert_own_property: expected property "sec-fetch-site" missing
FAIL sec-fetch-site - HTTPS downgrade-upgrade, no init assert_own_property: expected property "sec-fetch-site" missing
